+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 #   CHANGELOG
 
 ## 3/29/2025
-- 
+- able to pull table rows with date of word used and word itself
 
 ## 3/28/2025
 - Guesses function almost done, returned words that did NOT contain all characters that user identified are in word